+++ app/src/main/java/com/github/jan222ik/floodit/logic/FloodItGameState.kt
@@ -1,0 +1,5 @@
+package com.github.jan222ik.floodit.logic
+
+enum class FloodItGameState {
+    LOADING, PLACE_SOURCE, RUNNING, FINISHED
+}
