+++ app/src/main/java/com/github/jan222ik/floodit/logic/Point.kt
@@ -1,0 +1,3 @@
+package com.github.jan222ik.floodit.logic
+
+data class Point(val x: Int, val y: Int)
