+++ app/src/main/java/com/github/jan222ik/floodit/ui/components/CellSlider.kt
@@ -1,0 +1,41 @@
+package com.github.jan222ik.floodit.ui.components
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material.Slider
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import kotlin.math.roundToInt
+
+@Composable
+fun CellSlider(
+    title: (Int) -> String,
+    value: Int,
+    onValueChange: (Int) -> Unit,
+    valueRange: ClosedFloatingPointRange<Float>
+) {
+    Text(text = title(value))
+    Row(
+        modifier = Modifier.fillMaxWidth(),
+        horizontalArrangement = Arrangement.spacedBy(4.dp),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Text(
+            text = valueRange.start.toInt().toString(),
+        )
+        Slider(
+            modifier = Modifier.weight(1f, false),
+            value = value.toFloat(),
+            onValueChange = { onValueChange(it.roundToInt()) },
+            valueRange = valueRange,
+            steps = (valueRange.endInclusive - valueRange.start).dec().toInt()
+        )
+        Text(
+            text = valueRange.endInclusive.toInt().toString(),
+        )
+    }
+}
