+++ app/src/main/java/com/github/jan222ik/floodit/ui/components/ExpandingInfoChip.kt
@@ -1,0 +1,52 @@
+package com.github.jan222ik.floodit.ui.components
+
+import androidx.compose.animation.*
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.Icon
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Surface
+import androidx.compose.material.Text
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.outlined.Info
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@ExperimentalAnimationApi
+@Composable
+fun ExpandingInfoChip(text: String, isExpanded: Boolean, onClick: () -> Unit) {
+    val iS = remember {
+        MutableInteractionSource()
+    }
+    Surface(
+        shape = RoundedCornerShape(8.dp),
+        color = if (isExpanded) MaterialTheme.colors.surface else MaterialTheme.colors.background
+    ) {
+        Row(
+            modifier = Modifier
+                .padding(8.dp)
+                .clickable(interactionSource = iS, indication = null, onClick = onClick),
+            horizontalArrangement = Arrangement.spacedBy(8.dp),
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Icon(
+                imageVector = Icons.Outlined.Info,
+                contentDescription = null
+            )
+            androidx.compose.animation.AnimatedVisibility(
+                visible = isExpanded,
+                enter = fadeIn() + expandIn(expandFrom = Alignment.CenterStart),
+                exit = shrinkOut(shrinkTowards = Alignment.CenterStart) + fadeOut()
+            ) {
+                Text(text = text)
+            }
+        }
+    }
+}
